+++ js/main.js
@@ -1,3 +1,4 @@
 import './function.js';
 import './data.js';
+import './pedestal.js';
